Refactors identity handling in authenticator
Simplifies the identity wakeup process by removing the token parameter from the `getIdentity` method and introducing a `findSession` method to locate the storage entity by token.
This streamlines the logic and improves maintainability.
The authentication token is now set to identity after it is retrieved, instead of passing it as a parameter.
Adds function to find Identity by token.